# Improve SFTP performance on medium/high latency connections