@darkoperator 2021-08-30
Avery nice performance boost! This has been a complaint for a long time

> On Aug 30, 2021, at 3:41 PM, zybexXL ***@***.***> wrote:
> 
> 
> @zybexXL commented on this pull request.
> 
> In src/Renci.SshNet/ServiceFactory.cs <https://github.com/sshnet/SSH.NET/pull/866#discussion_r698757055>:
> 
> > @@ -132,7 +132,7 @@ public ISftpFileReader CreateSftpFileReader(string fileName, ISftpSession sftpSe
>              {
>                  var fileAttributes = sftpSession.EndLStat(statAsyncResult);
>                  fileSize = fileAttributes.Size;
> -                maxPendingReads = Math.Min(10, (int) Math.Ceiling((double) fileAttributes.Size / chunkSize) + 1);
> +                maxPendingReads = Math.Min(100, (int)Math.Ceiling((double)fileAttributes.Size / chunkSize) + 1);
> this line is also changed on PR #864 <https://github.com/sshnet/SSH.NET/pull/864> - here <https://github.com/sshnet/SSH.NET/pull/864/files#diff-abfc42015c3bccaae3e2012df18a1b2a24d66c6effb9ef5e30cbe6516023f967R135>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub <https://github.com/sshnet/SSH.NET/pull/866#pullrequestreview-742003951>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AAD7IHSQNLHAOK4ITGGLEWDT7PNHPANCNFSM5DCO2DGA>.
> Triage notifications on the go with GitHub Mobile for iOS <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675> or Android <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>. 
>