@zybexXL 2021-12-15
@drieseng, according to Wireshark, psftp sends **128 requests** in about 6ms. Data starts arriving 30ms later from my remote test server, and then psftp sends a new requests every ~32KB of received data.