@jjxtra 2022-07-20
Code doesn't compile for me, complaining about target frameworks being empty