@zybexXL 2022-07-20
I just pulled the branch and compiled in VS2019. This code does not touch any Framework/Project config file.
There are compile warnings about net.core 2.1/2.2/3.0 being deprecated, but that comes from the main line and can be ignored for now.